Hook up routes for the homepage and notfound
Basic routing is implemented with this commit. Top-level pages are in
the /pages folder, and links are created using the new Link component
(which navs via history instead of the browser).

Then, it's a simple update to the <app> component to re-render based on
the page to the correct component